requirements.txt
Pillow==9.4.0
python-dotenv==1.0.0
transformers==4.35.2
langchain==0.0.352
streamlit==1.29.0
requests
